refactor(schemas): move validation tests into internal/jsonschema sub-module
Keeps santhosh-tekuri/jsonschema/v6 out of rig's main go.mod — the
library is a test-only concern and belongs in the isolated sub-module
alongside the schema generator.
