+++ internal/jsonschema/go.mod
@@ -5,6 +5,7 @@ go 1.26
 require (
 	github.com/invopop/jsonschema v0.13.0
 	github.com/k0sproject/rig v0.21.0
+	github.com/santhosh-tekuri/jsonschema/v6 v6.0.2
 	gopkg.in/yaml.v3 v3.0.1
 )
